Add support for gitlab auth/download
We only support PAT auth for gitlab, since the GCM built-in oauth cred does not have the scope we need (read_api) for archive downloading.

Added the boilerplate to program.cs so that the GCM UI can properly ask for credentials when downloading from private repositories. For GH, we don't use namespaced creds since we don't need any special scopes, so we reuse the default namespace used by the regular git operations.